Refactor regarding code review
