+++ src/main/java/com/jvm_bloggers/utils/NowProvider.java
@@ -6,9 +6,6 @@ import java.time.ZoneId;
 
 public interface NowProvider {
 
-    String DEFAULT_ZONE_NAME = "Europe/Warsaw";
-    ZoneId DEFAULT_ZONE = ZoneId.of(DEFAULT_ZONE_NAME);
-
     LocalDateTime now();
 
     LocalDate today();
