+++ src/main/java/com/jvm_bloggers/utils/ZoneTimeProvider.java
@@ -4,9 +4,13 @@ import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.ZoneId;
 
 @Component
 public final class ZoneTimeProvider implements NowProvider {
+
+    public static final String DEFAULT_ZONE_NAME = "Europe/Warsaw";
+    public static final ZoneId DEFAULT_ZONE = ZoneId.of(DEFAULT_ZONE_NAME);
 
     @Override
     public LocalDateTime now() {
