+++ src/test/groovy/com/jvm_bloggers/core/data_fetching/blog_posts/RssCheckingActorSpec.groovy
@@ -7,7 +7,7 @@ import akka.testkit.JavaTestKit
 import com.jvm_bloggers.core.rss.SyndFeedProducer
 import com.jvm_bloggers.entities.blog.Blog
 import com.jvm_bloggers.entities.blog.BlogType
-import com.jvm_bloggers.utils.NowProvider
+import com.jvm_bloggers.utils.ZoneTimeProvider
 import com.rometools.rome.feed.synd.SyndEntry
 import com.rometools.rome.feed.synd.SyndFeed
 import io.vavr.control.Option
@@ -25,7 +25,7 @@ class RssCheckingActorSpec extends Specification {
         .author('Tomasz Dziurko')
         .rss('http://tomaszdziurko.pl/feed/')
         .url('url')
-        .dateAdded(LocalDateTime.now(NowProvider.DEFAULT_ZONE))
+        .dateAdded(new ZoneTimeProvider().now())
         .blogType(BlogType.PERSONAL)
         .build()
